pyproject.toml
[project]
name = "vulnerable-test"
version = "0.1.0"
description = "Test project with known vulnerable dependencies for security testing"
dependencies = [
    "urllib3==1.26.0",
    "cryptography==43.0.3",
    "pillow==8.2.0",
    "jinja2==2.11.3",
    "pyyaml==5.4.1",
    "requests==2.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

